refactor(cli): moving wallet mod into its own mod folder
