+++ sn_cli/src/subcommands/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2023 MaidSafe.net limited.
+// Copyright 2024 MaidSafe.net limited.
 //
 // This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
 // Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
@@ -16,7 +16,7 @@ use clap::Subcommand;
 pub(super) enum SubCmd {
     #[clap(name = "wallet", subcommand)]
     /// Commands for wallet management
-    Wallet(wallet::WalletCmds),
+    Wallet(wallet::hot_wallet::WalletCmds),
     #[clap(name = "files", subcommand)]
     /// Commands for file management
     Files(files::FilesCmds),
